Move component scripts into component library
