+++ components/library/filter/filter.js
@@ -1,7 +1,8 @@
+// Component script: Filter
 import noUiSlider from 'nouislider';
 import flatpickr from 'flatpickr/dist/flatpickr.js';
 import rangePlugin from 'flatpickr/dist/plugins/rangePlugin.js';
-import { swiperInit } from '../plugins/swiperInit.js';
+import { swiperInit } from '../swiper/swiperInit.js';
 
 export function filter() {
 	const filterForm = document.querySelector('[data-filter-form]');
